+++ packable/packable/CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.6.0 - 2022-08-23
 
+### Added
+
+- `Packable::UnpackVisitor` associated type;
+- `Packable::{unpack, unpack_verified}` argument `visitor`;
+- `#[packable(unpack_visitor = _)]` attribute;
+
 ### Changed
 
 - Updated dependencies;